More efficient render sorting

+++ src/main/core/graphic/BitMap.java
@@ -5,7 +5,7 @@ import org.lwjgl.system.NativeResource;
 
 import java.lang.foreign.MemorySegment;
 
-public record BitMap(int width, int height, MemorySegment data, int glFormat, int glType) implements NativeResource {
+public record BitMap(short width, short height, MemorySegment data, int glFormat, int glType) implements NativeResource {
     @Override
     public void free() {
         MemoryUtil.nmemFree(data.address());
